Add updateSubscriptionEndsAt method and enhance SubscriptionBuilder for trial management

+++ src/main/java/com/oixan/stripecashier/entity/Subscription.java
@@ -38,7 +38,7 @@ public class Subscription {
     private String stripePrice;
 
     @Column
-    private Integer quantity;
+    private Long quantity;
 
     @Column(name = "trial_ends_at")
     private LocalDateTime trialEndsAt;
@@ -102,11 +102,11 @@ public class Subscription {
         this.stripePrice = stripePrice;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
